requirements.txt
openai==1.37
streamlit==1.34
tiktoken==0.7
pymongo==4.7
pinecone-client==5.0
pinecone-text==0.9